add extension to UIView to simplify adding and removing subviews
